fix supplier deposit

+++ netforce_account/netforce_account/models/account_invoice.py
@@ -37,7 +37,7 @@ class Invoice(Model):
     _multi_company = True
     _fields = {
         "type": fields.Selection([["out", "Receivable"], ["in", "Payable"]], "Type", required=True),
-        "inv_type": fields.Selection([["invoice", "Invoice"], ["credit", "Credit Note"], ["debit", "Debit Note"], ["prepay", "Prepayment"], ["overpay", "Overpayment"]], "Subtype", required=True, search=True),
+        "inv_type": fields.Selection([["invoice", "Invoice"], ["credit", "Credit Note"], ["debit", "Debit Note"]], "Subtype", required=True, search=True),
         "number": fields.Char("Number", search=True),
         "ref": fields.Char("Ref", size=256, search=True),
         "memo": fields.Char("Memo", size=1024, search=True),
@@ -529,6 +529,8 @@ class Invoice(Model):
         obj = self.browse(ids)[0]
         if obj.state != "waiting_payment":
             raise Exception("Invalid status")
+        if obj.payment_entries:
+            raise Exception("There are still payment entries for this invoice")
         if obj.move_id:
             obj.move_id.void()
             obj.move_id.delete()
@@ -570,7 +572,7 @@ class Invoice(Model):
             paid_amt = 0
             for pmt in inv.payment_entries:
                 if inv.type == "in":
-                    paid_amt -= pmt.debit # TODO: currency
+                    paid_amt += pmt.debit # TODO: currency
                 else:
                     paid_amt += pmt.credit # TODO: currency
             vals["amount_paid"] = paid_amt
